start monte carlo
