+++ game-engine/build_python.sh
@@ -3,6 +3,6 @@ echo "Building project"
 cargo build --release
 
 cp target/release/stratego-exec game-stratego
-cp target/release/libstratego_lib.so ai-python
-mv ai-python/libstratego_lib.so ai-python/stratego_engine.so
+cp target/release/libstratego_lib.so ai_python
+mv ai_python/libstratego_lib.so ai_python/stratego_engine.so
